requirements.txt
langchain==0.0.312
faiss-cpu==1.7.4
sentence-transformers==2.2.2
streamlit==1.25.0
openai==0.27.10
python-dotenv==1.0.0